requirements.txt
# Core data processing
pandas==2.1.4
numpy==1.26.2

# Database
psycopg2-binary==2.9.9

# File watching
watchdog==3.0.0

# Visualization
streamlit==1.29.0
plotly==5.18.0
